FN-DSA (Falcon): Falcon API headers moved to a dedicated public API level; minor coding-guideline and robustness cleanups; further adjustments and justifications based on static code analysis findings. HQC: Added dedicated HQC API headers; minor coding-guideline related cleanups; additional adjustments based on static code analysis findings.

+++ Hqc/Hqc128/FsmSw_Hqc128_fft.h
@@ -65,7 +65,7 @@
 /* PUBLIC FUNCTION PROTOTYPES                                                                                         */
 /**********************************************************************************************************************/
 
-void FsmSw_Hqc128_Fft(uint16 *const w, const uint16 *const f, uint8 f_coeffs);
+void FsmSw_Hqc128_Fft(uint16 *const w, const uint16 *const f, uint16 f_coeffs);
 
 void FsmSw_Hqc128_Fft_Retrieve_Error_Poly(uint8 *const err, const uint16 *const w);
