+++ Hqc/Hqc256/FsmSw_Hqc256_fft.h
@@ -63,7 +63,7 @@
 /* PUBLIC FUNCTION PROTOTYPES                                                                                         */
 /**********************************************************************************************************************/
 
-void FsmSw_Hqc256_Fft(uint16 *const w, const uint16 *const f, uint32 f_coeffs);
+void FsmSw_Hqc256_Fft(uint16 *const w, const uint16 *const f, uint16 f_coeffs);
 
 void FsmSw_Hqc256_Fft_Retrieve_Error_Poly(uint8 *const err, const uint16 *const w);
